Aerospace_Incident_Report_Analyzer/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astroid==3.2.2
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dill==0.3.8
distro==1.9.0
faiss-cpu==1.11.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
mccabe==0.7.0
narwhals==1.39.1
numpy==2.0.0
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.2
pandas_access==0.0.1
pillow==11.2.1
platformdirs==4.2.2
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pylint==3.2.5
pylit==0.8.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.0
setuptools==70.1.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tomlkit==0.12.5
tornado==6.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.4.0
watchdog==6.0.0
wheel==0.43.0
zstandard==0.23.0
